requirements_dev.txt
django-debug-toolbar==3.7.0
black==22.10.0
flake8==6.0.0
